feat: add issue to github organisation project board
it is not possible to triage an issue to an organization board with just
the normal github actions. we need different permissions.
and rather than creating a specific user an app makes more sense
this app solves this problem of triaging an issue to an org board

specify the organization and the board name as well as the column to
triage to in the config
create an `triaging.yml` inside your `.github` folder of the
repositories that would use this app